Don't fail if testing workspace matches workspace in environment (#10)
* This change fixes a bug that can occurr when the workspace used for
testing is the same as the workspace configured in the environment.

This also adds more test cases and bette CI

* Minor formatting fixes

* Remove samples unit testing

+++ unit/unit.go
@@ -55,7 +55,9 @@ func RunUnitTests(fixture *UnitTestFixture) {
 		terraform.FormatArgs(fixture.TfOptions, "workspace", "show")...)
 
 	terraform.WorkspaceSelectOrNew(fixture.GoTest, fixture.TfOptions, workspaceName)
-	defer terraform.RunTerraformCommand(fixture.GoTest, fixture.TfOptions, "workspace", "delete", workspaceName)
+	if startingWorkspaceName != workspaceName {
+		defer terraform.RunTerraformCommand(fixture.GoTest, fixture.TfOptions, "workspace", "delete", workspaceName)
+	}
 	defer terraform.WorkspaceSelectOrNew(fixture.GoTest, fixture.TfOptions, startingWorkspaceName)
 
 	tfPlanFilePath := filepath.FromSlash(fmt.Sprintf("%s/%s.plan", os.TempDir(), random.UniqueId()))
